requirements.txt
jupyterlab==4.3.5
numpy==2.2.2
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2
scipy==1.15.1
scikit-learn==1.6.1
pyspark==3.5.5
xgboost==2.1.3
apache-airflow==2.10.4
python-dateutil==2.9.0
pyarrow==17.0.0
# Optional alternative parquet engine:
# fastparquet==2024.5.0